Map api endpoint to list reservations list
