+++ src/App.js
@@ -1,9 +1,5 @@
 import React from 'react';
-import {
-  BrowserRouter,
-  Routes,
-  Route,
-} from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
 import './App.css';
 import Sidebar from './components/sidebar/Sidebar';
 import Home from './components/home/Home';
@@ -11,11 +7,11 @@ import Login from './components/loginpage/Login';
 import AddDcotor from './components/addDoctor/AddDoctor';
 import DelDoctor from './components/DelDoctor/DelDoctor';
 import Reserv from './components/reserve/Reserv';
-import Reservation from './components/reservation/Reservation';
+import Reservations from './components/reservation/Reservations';
 
 function App() {
   return (
-    <BrowserRouter>
+    <div className="app">
       <Sidebar />
       <Routes>
         <Route index element={<Login />} />
@@ -23,9 +19,9 @@ function App() {
         <Route path="adddoctor" element={<AddDcotor />} />
         <Route path="deldoctor" element={<DelDoctor />} />
         <Route path="reserveform" element={<Reserv />} />
-        <Route path="myreservations" element={<Reservation />} />
+        <Route path="myreservations" element={<Reservations />} />
       </Routes>
-    </BrowserRouter>
+    </div>
   );
 }
