--- src/components/reservation/Reservation.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Reservation = () => (
-  <div>
-    <h1>Reservation Page</h1>
-  </div>
-);
-
-export default Reservation;
