+++ src/components/reservation/reservation.css
@@ -1,3 +1,19 @@
 body {
   padding: 0;
 }
+
+.reservation-page {
+  flex-direction: column;
+  width: 100%;
+  justify-content: center;
+  padding: 1rem;
+  gap: 2rem;
+}
+
+.reservation-info {
+  flex-direction: column;
+  justify-content: space-between;
+  padding: 1rem;
+  align-items: center;
+  gap: 1rem;
+}
